+++ src/users/pages/Members.vue
@@ -76,6 +76,7 @@
       :group="group"
       :sorting="sorting"
       @create-trust="createTrust"
+      @revoke-trust="revokeTrust"
     />
   </QCard>
 </template>
@@ -130,6 +131,7 @@ export default {
   methods: {
     ...mapActions({
       createTrust: 'currentGroup/trustUser',
+      revokeTrust: 'currentGroup/revokeTrust',
     }),
     toggleSorting () {
       if (this.sorting === 'joinDate') {